update confirm address routes in protected renew app
