+++ frontend/app/page-ids.ts
@@ -17,8 +17,8 @@ export const pageIds = {
       parentOrGuardianRequired: 'CDCP-PROT-RENW-0010',
       confirmPhone: 'CDCP-PROT-RENW-0011',
       confirmEmail: 'CDCP-PROT-RENW-0012',
-      confirmAddress: 'CDCP-PROT-RENW-0013',
-      updateAddress: 'CDCP-PROT-RENW-0014',
+      confirmHomeAddress: 'CDCP-PROT-RENW-0013',
+      confirmMailingAddress: 'CDCP-PROT-RENW-0014',
       confirmCommunicationPreference: 'CDCP-PROT-RENW-0015',
       confirmFederalProvincialTerritorialBenefits: 'CDCP-PROT-RENW-0016',
       reviewAdultInformation: 'CDCP-PROT-RENW-0017',
