+++ frontend/public/locales/en/protected-renew.json
@@ -328,37 +328,20 @@
       "confirm-email-valid": "Enter confirmation of email address in the correct format, such as name@example.com"
     }
   },
-  "confirm-address": {
-    "page-title": "Address",
-    "have-you-moved": "Have you moved or changed your mailing address since you applied for the Canadian Dental Care Plan?",
-    "is-home-address-same-as-mailing-address": "Is your home address the same as your mailing address?",
-    "radio-options": {
-      "yes": "Yes",
-      "no": "No"
-    },
-    "help-message": "The address we have on file can also be found on the renewal letter you received from Service Canada.",
-    "error-message": {
-      "has-address-changed-required": "Select whether you have moved or changed your mailing address",
-      "is-home-address-same-as-mailing-address-required": "Select whether or not your mailing address and home address are the same"
-    },
-    "back-btn": "Back",
-    "continue-btn": "Continue",
-    "cancel-btn": "Cancel",
-    "save-btn": "Save"
-  },
   "update-address": {
-    "page-title": "Update address",
     "mailing-address": {
+      "page-title": "Mailing address",
       "header": "Mailing address"
     },
     "home-address": {
+      "page-title": "Home address",
       "header": "Home address",
       "use-mailing-address": "My home address is the same as my mailing address"
     },
     "address-field": {
       "address": "Address",
       "mailing-address": "Mailing address",
-      "address-note": "Include apartment number (if applicable), street number, street name. For example: 4B - 123 Main St.",
+      "address-note": "Include apartment number (if applicable), street number, street name. For example: 123 Main Street Suite 4B",
       "country": "Country",
       "province": "Province, territory, state, or region",
       "city": "City or town",
@@ -372,24 +355,24 @@
     "save-btn": "Save",
     "error-message": {
       "mailing-address": {
-        "address-required": "Enter mailing address, typically number and street",
-        "city-required": "Enter a city or town for your mailing address",
-        "country-required": "Select a country for your mailing address",
-        "postal-code-required": "Enter postal code or zip code for your mailing address",
-        "postal-code-valid": "Enter mailing address postal code in the correct format, such as A1A 1A1",
-        "province-required": "Select a province, territory, state or region for your mailing address",
-        "zip-code-valid": "Enter mailing address zip code in the correct format, such as 12345 or 12345-6789",
+        "address-required": "Address must be provided",
+        "city-required": "City must be provided",
+        "country-required": "Country must be selected",
+        "postal-code-required": "Postal code or Zip code must be provided",
+        "postal-code-valid": "Invalid postal code (A1A 1A1)",
+        "province-required": "Province must be selected",
+        "zip-code-valid": "Invalid zip code (12345)",
         "invalid-postal-code-for-country": "A Canadian postal code in the mailing address was entered but Canada was not selected as the country. Please select Canada, and the proper province, or change the postal code",
         "invalid-postal-code-for-province": "The postal code does not match the province or territory you selected"
       },
       "home-address": {
-        "address-required": "Enter home address, typically number and street",
-        "city-required": "Enter a city or town for your home address",
-        "country-required": "Select a country for your home address",
-        "postal-code-required": "Enter postal code or zip code for your home address",
-        "postal-code-valid": "Enter home address postal code in the correct format, such as A1A 1A1",
-        "province-required": "Select a province, territory, state or region for your home address",
-        "zip-code-valid": "Enter home address zip code in the correct format, such as 12345 or 12345-6789",
+        "address-required": "Address must be provided",
+        "city-required": "City must be provided",
+        "country-required": "Country must be selected",
+        "postal-code-required": "Postal code or Zip code must be provided",
+        "postal-code-valid": "Invalid postal code (A1A 1A1)",
+        "province-required": "Province must be selected",
+        "zip-code-valid": "Invalid zip code (12345)",
         "invalid-postal-code-for-country": "A Canadian postal code in the home address was entered but Canada was not selected as the country. Please select Canada, and the proper province, or change the postal code",
         "invalid-postal-code-for-province": "The postal code does not match the province or territory you selected"
       },
